ReadableBufferExtensions.GetUInt32 now works over sequences of ReadOnlyMemory<T>
In addition, I optimized the method a bit. It was heap allocating in cases where the first segment was too short
to pars an int and the second was too long for them both to fit in 128 bytes.
Now, the algorithm is "optimistic". It tries to parse without allocating, and only allocates when it fails.